made store variable 8 bit wide

+++ firmware/application/hw/encoder.hpp
@@ -32,7 +32,7 @@ class Encoder {
 
    private:
     uint_fast8_t state{0};
-    uint_fast16_t store{0};
+    uint_fast8_t store{0};
 };
 
 #endif /*__ENCODER_H__*/
